IOS-2973 Adapt CommonExplorerURLService to BlockchainSdk
